+++ web/babel.config.js
@@ -1,11 +1,12 @@
 module.exports = (api) => {
-  api.cache(true);
+  api.cache(false);
   return {
     presets: ["next/babel"],
     plugins: [
       [
         "tailwind-factory/plugin",
         {
+          logs: "debug",
           styles: {
             config: require("./tailwind.config"),
             outputPath: "src/styles/generated.css"