+++ web/components/Code/styles.ts
@@ -1,0 +1,10 @@
+import { tf } from "tailwind-factory";
+
+export const CodeContainer = tf("pre", `
+  bg-slate-900
+  p-8
+  rounded-lg
+  text-lg
+  text-slate-50
+  min-w-[calc(20rem)]
+`);