+++ web/src/pages/index.tsx
@@ -1,8 +1,10 @@
 import { Container } from "@/styles";
+import { Button } from "components/Button";
+import { Code } from "components/Code";
 import Image from "next/image";
 
 export default function Home() {
-  return <Container className="2xl:text-red-500">
+  return <Container>
     <Image
       src="/logo.png"
       width={1024}
@@ -10,5 +12,27 @@ export default function Home() {
       alt="Library logo, the text was made with two colors: purple and blue. It says 'tailwind factory'"
     />
     <div className="background-left"/>
+    <div className="background-right"/>
+    <div id="button-group">
+      <a tabIndex={-1} href="https://github.com/L-Marcel/tailwind-factory">
+        <Button>
+          Get Started
+        </Button>
+      </a>
+      <Button theme="secondary">
+        Next Example
+      </Button>
+    </div>
+    <section>
+      <Code>
+        <span>flex</span>
+        <span>flex-col</span>
+        <button>test</button>
+      </Code>
+      <Code>
+        <span>flex</span>
+        <span>flex-col</span>
+      </Code>
+    </section>
   </Container>;
 }